@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 2.0.0+16
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -42,6 +42,8 @@ dependencies:
   json_annotation: ^4.3.0
   in_app_review: ^2.0.3
   shared_preferences: ^2.0.11
+  purchases_flutter: ^3.0.0
+  cached_network_image: ^3.2.0
 
 dev_dependencies:
   flutter_test:
